refactor: move definition of tidy_response from vertex.py to data_parsing.py
